[DEV-11187] Creates unit tests and adds pydocs for _structure_file_reponse. Renames helper file
