+++ usaspending_api/download/tests/unit/test_list_unlinked_awards_downloads.py
@@ -1,7 +1,7 @@
 import datetime
 from unittest.mock import MagicMock
 
-from usaspending_api.download.helpers.latest_download_file_helpers import get_last_modified_file
+from usaspending_api.download.helpers.download_file_helpers import get_last_modified_file
 
 
 def test_get_last_modified_file():
