+++ usaspending_api/download/v2/list_unlinked_awards_downloads.py
@@ -3,7 +3,7 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from usaspending_api.common.cache_decorator import cache_response
-from usaspending_api.download.helpers.latest_download_file_helpers import (
+from usaspending_api.download.helpers.download_file_helpers import (
     get_last_modified_download_file_by_prefix,
     remove_file_prefix_if_exists,
 )
